split instance implementation from constructor
