+++ src/Constraint.php
@@ -52,6 +52,19 @@ final class Constraint
     }
 
     /**
+     * @template A of object
+     * @psalm-pure
+     *
+     * @param class-string<A> $class
+     *
+     * @return self<mixed, A>
+     */
+    public static function instance(string $class): self
+    {
+        return new self(Constraint\Instance::of($class));
+    }
+
+    /**
      * @psalm-pure
      * @template A
      * @template B
